+++ src/test.cpp
@@ -884,12 +884,12 @@ void Test::formula() {
     OeisSequence seq(e.first);
     Log::get().info("Testing formula for " + seq.id_str() + ": " + e.second);
     auto p = parser.parse(seq.getProgramPath());
-    auto r = FormulaGenerator::generate(p, false);
-    if (!r.first) {
+    Formula f;
+    if (!FormulaGenerator::generate(p, seq.id, f, false, false)) {
       Log::get().error("Cannot generate formula from program", true);
     }
-    if (r.second.toString(false) != e.second) {
-      Log::get().error("Unexpected formula: " + r.second.toString(false), true);
+    if (f.toString(false) != e.second) {
+      Log::get().error("Unexpected formula: " + f.toString(false), true);
     }
   }
 }